fix types and refactor
